Leaf: use a 32 bit current accumulator
The 1 s current accumulator was int64_t with a uint32_t sample count. One second
of 10 ms 0x1DB frames at the signal's full scale reaches about 102,000, so 32 bits
has four orders of magnitude of headroom and the mean no longer needs the 64 bit
division helpers.

The count is made signed at the same time: int32_t sum + uint32_t count would
promote the rounding arithmetic in update_values() to unsigned and turn every
discharge window into a large positive current.

+++ Software/src/battery/NISSAN-LEAF-BATTERY.cpp
@@ -62,7 +62,7 @@ void NissanLeafBattery::
   if (battery_Current2_sample_count > 0) {
     // Calculate the mean directly in dA before publishing so the fractional
     // part of the raw 0.5 A samples is preserved. Round symmetrically around zero.
-    const int64_t current_dA_sum = battery_Current2_sum_raw * 5;
+    const int32_t current_dA_sum = battery_Current2_sum_raw * 5;
     if (current_dA_sum >= 0) {
       datalayer_battery->status.current_dA =
           (int16_t)((current_dA_sum + battery_Current2_sample_count / 2) / battery_Current2_sample_count);
